moved generateInteractionXDM and trackWithDataMethods to a internal class OptimizeTrackingUtils
